Profile page Viewing Ventures
Before this update, when the user "added a venture," the Profile page
would not display the user's venture, and would defaultly say: "You
aren't part of a venture!"

Now, with the changes, the Profile page detects all the newly added
Ventures that the user created.

For future: Need to add the Ventures that the User plays a role in, not
just the ventures that the User created.