All Existing Ventures page, Admin's profile
1) Allows all users to be able to see all existing ventures. Each user
can click on the Venture's admin and find the admin's profile correctly
now. Venture correctly finds admin.
2) Also changed the DataSource to persistent rather than create-drop.
This way we don't have to keep adding new ventures.
3) Fixed bunch of bugs

TODO: Started Roles on Profile page. The Profile page currently
displays only the Ventures that they've created. Need to first allow
Admins to be able to create filledRoles and availableRoles. Next, need
to allow users to fill in availableRoles. Lastly, the Profile page
should display for the current user, all the Roles that he/she is in
for each Venture.