Added attributes specific to Project.
Added the attributes specific to Project. Currently, the constraints
are there, but do not show to the user. If bad inputs inputted, the
Project is not created. Also, need to update the "EDIT" and add the
other pages with respect to Project. Found out that the
DataSource.groovy needs to use create-drop instead of update.